+++ packages/sdk/CHANGELOG.md
@@ -19,6 +19,7 @@
 - `parseTokenUpdatePayload`, which decodes the CBOR encoded operations and returns a corresponding payload with the
   operations decoded into `(TokenOperation | UnknownTokenOperation)[]`
 - `parseTokenModuleRejectReason`, which decodes `Cbor.Type` into `TokenModuleRejectReason | UnknownTokenModuleRejectReason`.
+- `CborContractAddress` which represents CIS-7 compatible contract addresses.
 
 ### Breaking changes